backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiohttp==3.9.1
python-multipart==0.0.6

# Postgres / Supabase
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
pandas>=2.1.0

# Conversational agent (aqi_agent package, local)
pydantic-ai-slim[google,groq,mcp,logfire]>=0.0.40
sse-starlette>=2.1.0
langfuse>=2.50.0
tavily-python>=0.5.0
httpx>=0.27.0
